==> Notebooks/02_scrape_clean.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
import requests
import numpy as np

nodeslist = pd.read_json('nodeslist.json')
nodesDF = nodeslist[~nodeslist['hasChild']]
nodesDF = nodesDF[nodesDF['value']!='']
nodesDF.shape

nodesDF.head()

home = 'https://www.seattle.gov/util/MyServices/WhereDoesItGo/'
my_resp = lambda x: requests.get(home+x+'/index.htm', timeout=5)
my_content = lambda x: BeautifulSoup(x.content, "html.parser")
my_h3 = lambda x: x.find_all('h3')[1:]
my_labels = lambda x: [s['class'][0] for s in x]

labels = []
for i in range(nodesDF.shape[0]):
    _resp = my_resp(nodesDF.iloc[i]['value'])
    _cont = my_content(_resp)
    _h3 = my_h3(_cont)
    try:
        _lab = my_labels(_h3)
        labels.append(_lab)
    except:
        labels.append([''])

compost = np.array([True if 'food' in s else False for s in labels])
garbage = np.array([True if 'garbage' in s else False for s in labels])
recycle = np.array([True if 'recycle' in s else False for s in labels])
binside = np.array([True if 'binSide' in s else False for s in labels])
donate  = np.array([True if 'donate' in s else False for s in labels])
transfer = np.array([True if 'transfer' in s else False for s in labels])

print(sum(compost)/len(compost))
print(sum(garbage)/len(compost))
print(sum(recycle)/len(compost))
print(sum(binside)/len(compost))
print(sum(donate)/len(compost))
print(sum(transfer)/len(compost))

compost2 = compost.copy()

for i in range(len(compost2)):
    _w = nodesDF['label'].iloc[i]
    _w = _w.lower()
    if (('plastic' in _w)|('glass' in _w)|('pot' in _w)
       |('jug' in _w)|('cup' in _w)|('bowl' in _w)
       |('spoon' in _w)|('fork' in _w)|('knife' in _w)
       |('spork' in _w)):
        compost2[i] = False

print(sum(compost2)/len(compost))

nodesDF['compost'] = compost2

nodesDF = nodesDF.drop(columns=['hasChild','value'])

nodesDF.head()

nodesDF.to_json('compost_or_not.json',orient='records')
