==> Notebooks/03_augment_clean.ipynb <==
import pandas as pd
import numpy as np
import pickle
from nltk.stem import WordNetLemmatizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction import stop_words

compostDF = pd.read_json('compost_or_not.json')

words = compostDF['label'].values
types = compostDF['compost'].values

extra_words = np.array(['Food Fruit and vegetables',
'Food, Bread, pasta, rice, and grains',
'Food, Eggshells and nutshells',
'Food, Coffee grounds and filters',
'Food, Tea bags',
'Food, Meat, fish, and chicken, pork, kaluas, carnitas, produce',
'Food, Dairy products like milk, butter, and cheese',
'Food, Bones and seafood shells, fish scale',
'Compostable (uncoated) paper such as paper napkins, paper towels, and pizza boxes',
'Both dry and wet pet food can be turned compost',
'Greasy or food-soiled pizza boxes go in the food and yard waste',
'Paper coffee filters are compostable and go in the food and yard waste together with coffee grounds',
'Dirty or clean uncoated (not shiny) paper take-out containers',
'Dirty or clean take-out containers labeled “Compostable” or “#7 PLA”',
'Greasy pizza boxes',
'Cardboard with grease or oil stains (like pizza boxes) go in the food and yard waste bin',
'Put cardboard that is mostly wet in the food and yard waste bin',
'Food, produce, vegetable',
'ingredients, spices',
'Natural foods, vegetable, fruits',
'Vegetable, produce',
'Vegetarian food',
'Plant, vegetable, food',
'whole food, whole grain, leaf vegetable',
'superfood, food, whole wheat'])

extra_types = np.repeat(True,len(extra_words))

words = np.append(words,extra_words)
types = np.append(types,extra_types)

_words = words[~types]
for i,word in enumerate(_words):
    _words[i] = word.replace('food','').replace('Food','')\
        .replace('vegetable','').replace('Vegetable','')\
        .replace('plant','').replace('Plant','')\
        .replace('meat','').replace('Meat','')
    
words[~types] = _words

labels = np.empty(len(types),dtype=int)
labels[types]=1
labels[~types]=0

idx = np.arange(0,len(labels))
np.random.shuffle(idx)
words,labels = words[idx],labels[idx]

pickle.dump((words,labels),open('words_labels.pkl','wb'))
